consolidate legacy workspace onto canonical backend
